Move DOMWidget out of widget.py
